requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
albumentations
opencv-python-headless
numpy
Pillow